+++ app/pages/dashboard/controller.js
@@ -1,0 +1,5 @@
+module.exports = app => {
+    app.controller('DashboardCtrl', function () {
+
+    });
+};
