--- app/pages/home/controller.js
@@ -1,5 +1,0 @@
-module.exports = app => {
-    app.controller('HomeCtrl', function () {
-
-    });
-};
