+++ public/index.html
@@ -13,7 +13,7 @@
 <body class="has-navbar-fixed-top">
     <nav class="navbar is-link is-fixed-top" role="navigation" aria-label="main navigation">
         <div class="navbar-brand">
-            <a class="navbar-item" data-ng-class="{ 'is-active': isActiveTab('home') }" href="/home">
+            <a class="navbar-item" data-ng-class="{ 'is-active': isActiveTab('dashboard') }" href="/dashboard">
                 <img src="./favicon.ico" alt="TBM">
             </a>
             <a class="navbar-item" data-ng-class="{ 'is-active': isActiveTab('tournaments') }" href="/tournaments">Tournaments</a>
